Добавил функцию состояния пользователя в корневой элемент
