+++ src/components/pages/CharInfoPage/CharInfoBlock.js
@@ -3,26 +3,18 @@ import SearchItem from "../../SearchItem/SearchItem";
 import { motion } from 'framer-motion'
 import { useNavigate } from "react-router-dom";
 import './CharInfoBlock.scss'
-import { useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { useEffect } from "react";
-import { onAuthStateChanged, getAuth } from 'firebase/auth'
-import { useronline } from "../LoginPage/LoginPageSlice";
+
 
 const CharInfoBlock = () => {
     const navigate = useNavigate()
-    const dispatch = useDispatch()
-
+    const userOnline = useSelector(state => state.login.userOnline)
     useEffect(() => {
-        const auth = getAuth()
-        onAuthStateChanged(auth, (user) => {
-            if (user !== null && user) {
-                dispatch(useronline(user.email))
-            } else {
-                dispatch(useronline('offline'))
-                navigate('/')
-            }
-        })
-    }, [dispatch, navigate])
+        if (!userOnline) {
+            navigate('/')
+        }
+    }, [userOnline, navigate])
 
     return (
         <motion.div
