+++ src/components/pages/FavoritePage/FavoritePage.js
@@ -4,24 +4,17 @@ import { motion } from 'framer-motion'
 import inkognito from '../../../images/inkognito.jpg'
 import { useNavigate } from "react-router-dom";
 import { useEffect } from 'react';
-import { onAuthStateChanged, getAuth } from 'firebase/auth'
-import { useronline } from "../LoginPage/LoginPageSlice";
-import { useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 
 const FavoritePage = () => {
     const navigate = useNavigate()
-    const dispatch = useDispatch()
+    const userOnline = useSelector(state => state.login.userOnline)
 
     useEffect(() => {
-        const auth = getAuth()
-        onAuthStateChanged(auth, (user) => {
-            if (user !== null && user) {
-                dispatch(useronline(user.email))
-            } else {
-                dispatch(useronline('offline'))
-            }
-        })
-    }, [dispatch, navigate])
+        if (!userOnline) {
+            navigate('/')
+        }
+    }, [userOnline, navigate])
 
 
     return (
