+++ src/components/pages/LoginPage/LoginPage.js
@@ -2,28 +2,12 @@ import Form from '../../Form/Form'
 import { motion } from 'framer-motion'
 import './LoginPage.scss'
 import { useNavigate } from 'react-router-dom'
-import { useDispatch, } from 'react-redux'
 import { useLazyLoginQuery } from '../../../apiFirebase/apiFireBaseSlice'
-import { useEffect } from 'react'
-import { onAuthStateChanged, getAuth } from 'firebase/auth'
-import { useronline } from './LoginPageSlice'
 
 
 const LoginPage = () => {
     const navigate = useNavigate()
-    const dispatch = useDispatch()
     const [triggerLoginUser] = useLazyLoginQuery()
-
-    useEffect(() => {
-        const auth = getAuth()
-        onAuthStateChanged(auth, (user) => {
-            if (user !== null && user) {
-                dispatch(useronline(user.email))
-            } else {
-                dispatch(useronline('offline'))
-            }
-        })
-    }, [dispatch])
 
     const onHandleSubmit = (args) => {
         triggerLoginUser(args)
@@ -33,7 +17,6 @@ const LoginPage = () => {
                 }
             })
     }
-
 
     return (
         <motion.div className='login'
