+++ src/components/pages/SearchPage/SearchPage.js
@@ -3,28 +3,20 @@ import { Link } from "react-router-dom";
 import './SearchPage.scss'
 import { motion } from 'framer-motion'
 import inkognito from '../../../images/inkognito.jpg'
-import { useDispatch } from 'react-redux';
 import { useNavigate } from "react-router-dom";
 import { useEffect } from 'react';
-import { useronline } from '../LoginPage/LoginPageSlice';
-import { onAuthStateChanged, getAuth } from 'firebase/auth'
+import { useSelector } from 'react-redux';
 
 
 const SearchPage = () => {
     const navigate = useNavigate()
-    const dispatch = useDispatch()
+    const userOnline = useSelector(state => state.login.userOnline)
 
     useEffect(() => {
-        const auth = getAuth()
-        onAuthStateChanged(auth, (user) => {
-            if (user !== null && user) {
-                dispatch(useronline(user.email))
-            } else {
-                dispatch(useronline('offline'))
-                navigate('/')
-            }
-        })
-    }, [dispatch, navigate])
+        if (!userOnline) {
+            navigate('/')
+        }
+    }, [userOnline, navigate])
 
     return (
         <motion.div className="app__search"
